fixing smells
